vulnanalizer/app/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
asyncpg==0.29.0
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
aiohttp==3.9.5
openpyxl==3.1.2
pydantic==2.5.0
requests==2.31.0
redis==5.0.1
PyJWT==2.8.0
psutil==5.9.6
passlib[bcrypt]==1.7.4
bcrypt==4.1.2 